# api/librebread.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for GetSmsParamsFormat.
const (
	Html GetSmsParamsFormat = "html"
	Json GetSmsParamsFormat = "json"
)

// LibreBreadSMSIds defines model for LibreBreadSMSIds.
type LibreBreadSMSIds = []string

// SMS defines model for SMS.
type SMS struct {
	From     string    `json:"From"`
	ID       string    `json:"ID"`
	Provider string    `json:"Provider"`
	Text     string    `json:"Text"`
	Time     time.Time `json:"Time"`
	To       string    `json:"To"`
}

// SMSList defines model for SMSList.
type SMSList = []SMS

// PostLibrePushJSONBody defines parameters for PostLibrePush.
type PostLibrePushJSONBody = map[string]interface{}

// PostLibreSendFormdataBody defines parameters for PostLibreSend.
type PostLibreSendFormdataBody struct {
	Data  *map[string]string `form:"Data,omitempty" json:"Data,omitempty"`
	ID    *int64             `form:"ID,omitempty" json:"ID,omitempty"`
	TTL   *int64             `form:"TTL,omitempty" json:"TTL,omitempty"`
	Text  *string            `form:"Text,omitempty" json:"Text,omitempty"`
	Title *string            `form:"Title,omitempty" json:"Title,omitempty"`
}

// GetSmsParams defines parameters for GetSms.
type GetSmsParams struct {
	// Limit The number of latest sms
	Limit *int64 `form:"limit,omitempty" json:"limit,omitempty"`

	// Format Output format
	Format *GetSmsParamsFormat `form:"format,omitempty" json:"format,omitempty"`
}

// GetSmsParamsFormat defines parameters for GetSms.
type GetSmsParamsFormat string

// PostLibrePushJSONRequestBody defines body for PostLibrePush for application/json ContentType.
type PostLibrePushJSONRequestBody = PostLibrePushJSONBody

// PostLibreSendFormdataRequestBody defines body for PostLibreSend for application/x-www-form-urlencoded ContentType.
type PostLibreSendFormdataRequestBody PostLibreSendFormdataBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /)
	Get(w http.ResponseWriter, r *http.Request)

	// (POST /libre/check)
	PostLibreCheck(w http.ResponseWriter, r *http.Request)

	// (POST /libre/push)
	PostLibrePush(w http.ResponseWriter, r *http.Request)
	// Send the fake SMS
	// (POST /libre/send)
	PostLibreSend(w http.ResponseWriter, r *http.Request)
	// Returns a list of SMS.
	// (GET /sms)
	GetSms(w http.ResponseWriter, r *http.Request, params GetSmsParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /)
func (_ Unimplemented) Get(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /libre/check)
func (_ Unimplemented) PostLibreCheck(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /libre/push)
func (_ Unimplemented) PostLibrePush(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Send the fake SMS
// (POST /libre/send)
func (_ Unimplemented) PostLibreSend(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Returns a list of SMS.
// (GET /sms)
func (_ Unimplemented) GetSms(w http.ResponseWriter, r *http.Request, params GetSmsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// Get operation middleware
func (siw *ServerInterfaceWrapper) Get(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Get(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostLibreCheck operation middleware
func (siw *ServerInterfaceWrapper) PostLibreCheck(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostLibreCheck(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostLibrePush operation middleware
func (siw *ServerInterfaceWrapper) PostLibrePush(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostLibrePush(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostLibreSend operation middleware
func (siw *ServerInterfaceWrapper) PostLibreSend(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostLibreSend(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSms operation middleware
func (siw *ServerInterfaceWrapper) GetSms(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetSmsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", r.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	// ------------- Optional query parameter "format" -------------

	err = runtime.BindQueryParameter("form", true, false, "format", r.URL.Query(), &params.Format)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "format", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSms(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/", wrapper.Get)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/libre/check", wrapper.PostLibreCheck)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/libre/push", wrapper.PostLibrePush)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/libre/send", wrapper.PostLibreSend)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/sms", wrapper.GetSms)
	})

	return r
}