# api/openapi_librebread.yml
openapi: 3.0.0
info:
  title: LibreBread API
  version: 2.0.0
paths:
  /:
    get:
      responses:
        '302':
          description: Redirect to default subsystem. SMS by default
          headers:
            Location:
              schema:
                type: string
  /sms:
    get:
      summary: Returns a list of SMS.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int64
            default: 50
          description: The number of latest sms
        - in: query
          name: format
          schema:
            type: string
            enum:
              - json
              - html
            default: html
          description: Output format
      responses:
        '200':
          description: A array of SMS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSList'
            text/html:
        '500':
          description: Internal error ocured
          content:
            text/plain:
              schema:
                type: string
  /libre/send:
    post:
      summary: Send the fake SMS
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - from
                - to
                - text
              properties:
                ID:
                  type: integer
                  format: int64
                Title:
                  type: string
                Text:
                  type: string
                Data:
                  type: object
                  additionalProperties:
                    type: string
                TTL:
                  type: integer
                  format: int64
      responses:
        '200':
          description: List of message ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibreBreadSMSIds'
        '400':
          description: Some filed is missed
        '500':
          description: Internal error ocured
          content:
            text/plain:
              schema:
                type: string
  /libre/check:
    post:
      description: Not implemented
      responses:
        '501':
          description: Not implemented
  /libre/push:
    post:
      description: Send push-like message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - to
                - text
      responses:
        '200':
          description: Push recived
        '400':
          description: Mailformed request body
        '422':
          description: Wrong input params
components:
  schemas:
    SMS:
      type: object
      required:
        - ID
        - Time
        - From
        - To
        - Text
        - Provider
      properties:
        ID:
          type: string
        Time:
          type: string
          format: date-time
        From:
          type: string
        To:
          type: string
        Text:
          type: string
        Provider:
          type: string
    SMSList:
      type: array
      items:
        $ref: '#/components/schemas/SMS'
    LibreBreadSMSIds:
      type: array
      items:
        type: string